move common logic out to a util
